Fix bug in #'adict--guess-dictionary-cons
The argument passed to #'adict-guess-dictionary-name must be the part
of the list that follows the first item of 'names (i.e., the sequence
of possible dictionaries for each language), not the entire list. If
that is not the case, #'adict-guess-dictionary-name considers the
two-letter language code (e.g., "en") to be another dictionary name.

If we define '("en" "american" "english") in #'adict-dictionary-list,
we expect the name of the dictionary to be "american" or "english", in
this order of preference, but never "en" — even if it happens to match
one of the values returned by #'ispell-valid-dictionary-list.